desktop-agent/python-agent/requirements.txt
# Python Agent with System Tray - Requirements

# Core dependencies
websockets>=12.0
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.1.0

# System tray support
pystray>=0.19.5

# Optional: For better performance
numpy>=1.24.0

# Build tools (for creating executables)
pyinstaller>=6.3.0
